+++ lib/src/dart/extensions/extensions.dart
@@ -1,3 +1,4 @@
+export 'datetime_extension.dart';
 export 'int_extension.dart';
 export 'list_extension.dart';
 export 'map_extension.dart';